Executable file uploaded
Bing.exe file adds itself to the windows startup and runs in the background forever. It checks for new bing daily wallpaper and gives a slideshow of the images present in the directory 'F:/Wallpapers'. If the directory not present it is created automatically.
Bing_daily.exe file checks for new bing daily wallpaper when executed.